Revert "Merge pull request #372 from aughey/betaflight"
This reverts commit a6a5f50ffbd470175e8dae9e17e4d5f4424996f9, reversing
changes made to 9cc55038516f52896ece099b6959a3413c25891f.
