+++ src/main/sensors/initialisation.c
@@ -214,7 +214,6 @@ const extiConfig_t *selectMPUIntExtiConfig(void)
 }
 
 #ifdef USE_FAKE_GYRO
-int16_t fake_gyro_values[XYZ_AXIS_COUNT] = { 0,0,0 };
 static void fakeGyroInit(uint16_t lpf)
 {
     UNUSED(lpf);
@@ -222,9 +221,7 @@ static void fakeGyroInit(uint16_t lpf)
 
 static bool fakeGyroRead(int16_t *gyroADC)
 {
-	for (int i = 0; i < XYZ_AXIS_COUNT; ++i) {
-		gyroADC[i] = fake_gyro_values[i];
-	}
+    memset(gyroADC, 0, sizeof(int16_t[XYZ_AXIS_COUNT]));
     return true;
 }
 
@@ -244,13 +241,9 @@ bool fakeGyroDetect(gyro_t *gyro)
 #endif
 
 #ifdef USE_FAKE_ACC
-int16_t fake_acc_values[XYZ_AXIS_COUNT] = {0,0,0};
-
 static void fakeAccInit(void) {}
 static bool fakeAccRead(int16_t *accData) {
-    for(int i=0;i<XYZ_AXIS_COUNT;++i) {
-		accData[i] = fake_acc_values[i];
-    }
+    memset(accData, 0, sizeof(int16_t[XYZ_AXIS_COUNT]));
     return true;
 }
